pyproject.toml
[tool.poetry]
name = "flask-server"
version = "0.1.0"
description = ""
authors = ["Angelo Aldecoa <angelokn14@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.0.3"
gunicorn = "^23.0.0"
flask-sqlalchemy = "^3.1.1"
argon2-cffi = "^23.1.0"
pyjwt = "^2.9.0"


[tool.poetry.group.dev.dependencies]
poethepoet = "^0.29.0"
pylint = "^3.3.1"
python-dotenv = "^1.0.1"
pytest = "^8.3.3"
pytest-flask = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
create_db = "python scripts/init_db.py"
dev = "flask --debug run"
test = "sh -c 'export FLASK_ENV=test && pytest'"
test-watch = "sh -c 'export FLASK_ENV=test && ptw'"
